Adding the layout element and refactor the index.tsx
