+++ apps/admin/pages/index.tsx
@@ -2,227 +2,173 @@ import {
   GraphStats,
   HotelStats,
   LatestClients,
-  NavBar,
-  SideBar,
-  SideBarLinkProps,
+  LayoutProtected,
 } from '@hotelbooking/shared-ui-component';
 import { Card, Grid, Row } from '@nextui-org/react';
-import {
-  AiFillBell,
-  AiFillHome,
-  AiFillSetting,
-  AiFillStar,
-  AiOutlineFileDone,
-  AiOutlineLogout,
-} from 'react-icons/ai';
-import { FaArrowRight, FaTelegram } from 'react-icons/fa';
+import { AiFillBell } from 'react-icons/ai';
+import { FaArrowRight } from 'react-icons/fa';
 
 export function Index() {
-  const sideBarUpLinks: SideBarLinkProps[] = [
-    {
-      href: '/',
-      iconElement: <AiFillHome className={'text-2xl text-teal-500'} />,
-      linkTitle: 'Dashboard',
-    },
-    {
-      href: '/',
-      iconElement: <AiFillStar className={'text-2xl text-teal-500'} />,
-      linkTitle: 'Favorite',
-    },
-    {
-      href: '/',
-      iconElement: <FaTelegram className={'text-2xl text-teal-500'} />,
-      linkTitle: 'Message',
-    },
-    {
-      href: '/',
-      iconElement: <AiOutlineFileDone className={'text-2xl text-teal-500 '} />,
-      linkTitle: 'Message',
-    },
-  ];
-  const sideBarDownLinks: SideBarLinkProps[] = [
-    {
-      href: '/',
-      iconElement: <AiFillSetting className={'text-2xl text-teal-500'} />,
-      linkTitle: 'Settings',
-    },
-    {
-      href: '/',
-      iconElement: <AiOutlineLogout className={'text-2xl text-teal-500 '} />,
-      linkTitle: 'Logout',
-    },
-  ];
   return (
-    <div className={'w-screen h-screen bg-gray-100 flex'}>
-      <SideBar
-        sideBarUpElements={sideBarUpLinks}
-        sideBarDownElements={sideBarDownLinks}
-      />
-      <main className={'flex-1 flex-col'}>
-        <NavBar />
-        <div className={'grid grid-cols-4 gap-12 pt-10 px-10'}>
-          <div className={'col-span-3 flex flex-col'}>
-            <h3
-              className={
-                'font-body font-black tracking-wider uppercase text-lg'
-              }
-            >
-              Dashboard
-            </h3>
-            <HotelStats />
-            <GraphStats />
-            <LatestClients />
-          </div>
-          <div className={''}>
-            <h3
-              className={
-                'font-body font-black tracking-wider uppercase text-lg'
-              }
-            >
-              Reports
-            </h3>
-            <Grid sm={12} md={12}>
-              <Card
-                css={{ borderRadius: '$xs', boxShadow: '$sm' }}
-                className={'transform hover:scale-105'}
-              >
-                <Card.Body css={{ py: '$4' }}>
-                  <div className={'flex items-center justify-around py-2'}>
-                    <AiFillBell
-                      className={
-                        'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
-                      }
-                    />
-                    <div className={'flex flex-col'}>
-                      <p className={'m-0 text-sm font-bold text-gray-500 '}>
-                        New reservation
-                      </p>
-                      <p
-                        className={
-                          'm-0 text-sm font-display font-medium text-gray-800'
-                        }
-                      >
-                        it was made from Axel Mwenze
-                      </p>
-                    </div>
-                  </div>
-                </Card.Body>
-                <Card.Footer>
-                  <Row justify="space-between">
-                    <p
-                      className={
-                        'text-gray-500 text-sm font-display text-gray-600'
-                      }
-                    >
-                      1 min ago
-                    </p>
-                    <a
-                      href="#"
-                      className={'text-teal-400 font-bold font-display flex'}
-                    >
-                      Details
-                      <FaArrowRight className={'ml-2'} />
-                    </a>
-                  </Row>
-                </Card.Footer>
-              </Card>
-            </Grid>
-            <Grid sm={12} md={12}>
-              <Card
-                css={{ borderRadius: '$xs', boxShadow: '$sm' }}
-                className={'transform hover:scale-105'}
-              >
-                <Card.Body css={{ py: '$4' }}>
-                  <div className={'flex items-center justify-around py-2'}>
-                    <AiFillBell
-                      className={
-                        'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
-                      }
-                    />
-                    <div className={'flex flex-col'}>
-                      <p className={'m-0 text-sm font-bold text-gray-500 '}>
-                        New reservation
-                      </p>
-                      <p
-                        className={
-                          'm-0 text-sm font-display font-medium text-gray-800'
-                        }
-                      >
-                        it was made from Axel Mwenze
-                      </p>
-                    </div>
-                  </div>
-                </Card.Body>
-                <Card.Footer>
-                  <Row justify="space-between">
-                    <p
-                      className={
-                        'text-gray-500 text-sm font-display text-gray-600'
-                      }
-                    >
-                      1 min ago
-                    </p>
-                    <a
-                      href="#"
-                      className={'text-teal-400 font-bold font-display flex'}
-                    >
-                      Details
-                      <FaArrowRight className={'ml-2'} />
-                    </a>
-                  </Row>
-                </Card.Footer>
-              </Card>
-            </Grid>
-            <Grid sm={12} md={12}>
-              <Card
-                css={{ borderRadius: '$xs', boxShadow: '$sm' }}
-                className={'transform hover:scale-105'}
-              >
-                <Card.Body css={{ py: '$4' }}>
-                  <div className={'flex items-center justify-around py-2'}>
-                    <AiFillBell
-                      className={
-                        'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
-                      }
-                    />
-                    <div className={'flex flex-col'}>
-                      <p className={'m-0 text-sm font-bold text-gray-500 '}>
-                        New reservation
-                      </p>
-                      <p
-                        className={
-                          'm-0 text-sm font-display font-medium text-gray-800'
-                        }
-                      >
-                        it was made from Axel Mwenze
-                      </p>
-                    </div>
-                  </div>
-                </Card.Body>
-                <Card.Footer>
-                  <Row justify="space-between">
-                    <p
-                      className={
-                        'text-gray-500 text-sm font-display text-gray-600'
-                      }
-                    >
-                      1 min ago
-                    </p>
-                    <a
-                      href="#"
-                      className={'text-teal-400 font-bold font-display flex'}
-                    >
-                      Details
-                      <FaArrowRight className={'ml-2'} />
-                    </a>
-                  </Row>
-                </Card.Footer>
-              </Card>
-            </Grid>
-          </div>
+    <LayoutProtected>
+      <div className={'grid grid-cols-4 gap-12 pt-10 px-10'}>
+        <div className={'col-span-3 flex flex-col'}>
+          <h3
+            className={'font-body font-black tracking-wider uppercase text-lg'}
+          >
+            Dashboard
+          </h3>
+          <HotelStats />
+          <GraphStats />
+          <LatestClients />
         </div>
-      </main>
-    </div>
+        <div className={''}>
+          <h3
+            className={'font-body font-black tracking-wider uppercase text-lg'}
+          >
+            Reports
+          </h3>
+          <Grid sm={12} md={12}>
+            <Card
+              css={{ borderRadius: '$xs', boxShadow: '$sm' }}
+              className={'transform hover:scale-105'}
+            >
+              <Card.Body css={{ py: '$4' }}>
+                <div className={'flex items-center justify-around py-2'}>
+                  <AiFillBell
+                    className={
+                      'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
+                    }
+                  />
+                  <div className={'flex flex-col'}>
+                    <p className={'m-0 text-sm font-bold text-gray-500 '}>
+                      New reservation
+                    </p>
+                    <p
+                      className={
+                        'm-0 text-sm font-display font-medium text-gray-800'
+                      }
+                    >
+                      it was made from Axel Mwenze
+                    </p>
+                  </div>
+                </div>
+              </Card.Body>
+              <Card.Footer>
+                <Row justify="space-between">
+                  <p
+                    className={
+                      'text-gray-500 text-sm font-display text-gray-600'
+                    }
+                  >
+                    1 min ago
+                  </p>
+                  <a
+                    href="#"
+                    className={'text-teal-400 font-bold font-display flex'}
+                  >
+                    Details
+                    <FaArrowRight className={'ml-2'} />
+                  </a>
+                </Row>
+              </Card.Footer>
+            </Card>
+          </Grid>
+          <Grid sm={12} md={12}>
+            <Card
+              css={{ borderRadius: '$xs', boxShadow: '$sm' }}
+              className={'transform hover:scale-105'}
+            >
+              <Card.Body css={{ py: '$4' }}>
+                <div className={'flex items-center justify-around py-2'}>
+                  <AiFillBell
+                    className={
+                      'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
+                    }
+                  />
+                  <div className={'flex flex-col'}>
+                    <p className={'m-0 text-sm font-bold text-gray-500 '}>
+                      New reservation
+                    </p>
+                    <p
+                      className={
+                        'm-0 text-sm font-display font-medium text-gray-800'
+                      }
+                    >
+                      it was made from Axel Mwenze
+                    </p>
+                  </div>
+                </div>
+              </Card.Body>
+              <Card.Footer>
+                <Row justify="space-between">
+                  <p
+                    className={
+                      'text-gray-500 text-sm font-display text-gray-600'
+                    }
+                  >
+                    1 min ago
+                  </p>
+                  <a
+                    href="#"
+                    className={'text-teal-400 font-bold font-display flex'}
+                  >
+                    Details
+                    <FaArrowRight className={'ml-2'} />
+                  </a>
+                </Row>
+              </Card.Footer>
+            </Card>
+          </Grid>
+          <Grid sm={12} md={12}>
+            <Card
+              css={{ borderRadius: '$xs', boxShadow: '$sm' }}
+              className={'transform hover:scale-105'}
+            >
+              <Card.Body css={{ py: '$4' }}>
+                <div className={'flex items-center justify-around py-2'}>
+                  <AiFillBell
+                    className={
+                      'transition-all duration-500 text-red-400 rounded-full  text-4xl transform hover:scale-110'
+                    }
+                  />
+                  <div className={'flex flex-col'}>
+                    <p className={'m-0 text-sm font-bold text-gray-500 '}>
+                      New reservation
+                    </p>
+                    <p
+                      className={
+                        'm-0 text-sm font-display font-medium text-gray-800'
+                      }
+                    >
+                      it was made from Axel Mwenze
+                    </p>
+                  </div>
+                </div>
+              </Card.Body>
+              <Card.Footer>
+                <Row justify="space-between">
+                  <p
+                    className={
+                      'text-gray-500 text-sm font-display text-gray-600'
+                    }
+                  >
+                    1 min ago
+                  </p>
+                  <a
+                    href="#"
+                    className={'text-teal-400 font-bold font-display flex'}
+                  >
+                    Details
+                    <FaArrowRight className={'ml-2'} />
+                  </a>
+                </Row>
+              </Card.Footer>
+            </Card>
+          </Grid>
+        </div>
+      </div>
+    </LayoutProtected>
   );
 }
