+++ libs/shared-ui-component/src/lib/layout-protected/layout-protected.module.scss
@@ -1,0 +1,7 @@
+/*
+ * Replace this with your own classes
+ *
+ * e.g.
+ * .container {
+ * }
+*/
