+++ libs/shared-ui-component/src/lib/layout-protected/layout-protected.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import LayoutProtected from './layout-protected';
+
+describe('LayoutProtected', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<LayoutProtected />);
+    expect(baseElement).toBeTruthy();
+  });
+});
